Refactored package structure into modular MVC structure.
