--- src/main/java/nl/tudelft/lifetiles/controller/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Contains the controllers.
- *
- * @author Joren Hammudoglu
- *
- */
-package nl.tudelft.lifetiles.controller;
-
