+++ src/main/java/nl/tudelft/lifetiles/data/controller/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.data.controller;
+
