+++ src/main/java/nl/tudelft/lifetiles/data/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * The data module. Select and display a (subset of) DNA strands.
+ *
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.data;
+
