+++ src/main/java/nl/tudelft/lifetiles/graph/controller/GraphController.java
@@ -1,25 +1,25 @@
-package nl.tudelft.lifetiles.controller;
+package nl.tudelft.lifetiles.graph.controller;
 
 import java.net.URL;
 import java.util.ResourceBundle;
 
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.layout.VBox;
+import javafx.scene.control.ScrollPane;
 
 /**
- * The controller of the main view.
+ * The controller of the graph view.
  *
  * @author Joren Hammudoglu
  *
  */
-public class MainController implements Initializable {
+public class GraphController implements Initializable {
 
     /**
      * The wrapper element.
      */
     @FXML
-    private VBox wrapper;
+    private ScrollPane wrapper;
 
     @Override
     public void initialize(final URL location, final ResourceBundle resources) {
