--- src/main/java/nl/tudelft/lifetiles/graph/jgrapht/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains all adapters required for the JGraphT library.
- */
-package nl.tudelft.lifetiles.graph.jgrapht;
