+++ src/main/java/nl/tudelft/lifetiles/graph/model/DefaultGraphParser.java
@@ -1,4 +1,4 @@
-package nl.tudelft.lifetiles.graph;
+package nl.tudelft.lifetiles.graph.model;
 
 import java.io.File;
 import java.io.IOException;
