+++ src/main/java/nl/tudelft/lifetiles/graph/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package contains all classes directly related to Graphs.
+ * The graph module. Visualizes the DNA alignment graph.
  */
 package nl.tudelft.lifetiles.graph;
+
