+++ src/main/java/nl/tudelft/lifetiles/lifetiles/controller/MainController.java
@@ -1,25 +1,25 @@
-package nl.tudelft.lifetiles.controller;
+package nl.tudelft.lifetiles.lifetiles.controller;
 
 import java.net.URL;
 import java.util.ResourceBundle;
 
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.ScrollPane;
+import javafx.scene.layout.VBox;
 
 /**
- * The controller of the graph view.
+ * The controller of the main view.
  *
  * @author Joren Hammudoglu
  *
  */
-public class GraphController implements Initializable {
+public class MainController implements Initializable {
 
     /**
      * The wrapper element.
      */
     @FXML
-    private ScrollPane wrapper;
+    private VBox wrapper;
 
     @Override
     public void initialize(final URL location, final ResourceBundle resources) {
