+++ src/main/java/nl/tudelft/lifetiles/lifetiles/model/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.lifetiles.model;
+
