+++ src/main/java/nl/tudelft/lifetiles/lifetiles/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * The LifeTiles core module.
+ *
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.lifetiles;
+
