+++ src/main/java/nl/tudelft/lifetiles/tree/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * The tree module. Visualizes the phylogenetic tree.
+ *
+ * @author Joren Hammudoglu
+ *
+ */
+package nl.tudelft.lifetiles.tree;
+
