+++ src/main/resources/fxml/GraphView.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.control.Label?>
 
 <ScrollPane xmlns:fx="http://javafx.com/fxml/1" prefHeight="342"
-	prefWidth="200.0" fx:controller="nl.tudelft.lifetiles.controller.GraphController"
+	prefWidth="200.0" fx:controller="nl.tudelft.lifetiles.graph.controller.GraphController"
 	id="graphPane">
 	<content>
 		<AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="342"
