+++ src/test/java/nl/tudelft/lifetiles/graph/jgrapht/GraphAdapterTest.java
@@ -5,11 +5,11 @@ import static org.junit.Assert.assertFalse;
 
 import java.util.Set;
 
-import nl.tudelft.lifetiles.graph.Edge;
-import nl.tudelft.lifetiles.graph.FactoryProducer;
-import nl.tudelft.lifetiles.graph.Graph;
-import nl.tudelft.lifetiles.graph.GraphFactory;
-import nl.tudelft.lifetiles.graph.SequenceSegment;
+import nl.tudelft.lifetiles.graph.model.Edge;
+import nl.tudelft.lifetiles.graph.model.FactoryProducer;
+import nl.tudelft.lifetiles.graph.model.Graph;
+import nl.tudelft.lifetiles.graph.model.GraphFactory;
+import nl.tudelft.lifetiles.graph.model.SequenceSegment;
 
 import org.junit.Before;
 import org.junit.BeforeClass;
